+++ dio/lib/src/transformers/sync_transformer.dart
@@ -6,6 +6,7 @@ import '../adapter.dart';
 import '../headers.dart';
 import '../options.dart';
 import '../transformer.dart';
+import '../utils.dart';
 
 /// The callback definition for decoding a JSON string.
 typedef JsonDecodeCallback = FutureOr<dynamic> Function(String);
@@ -27,11 +28,21 @@ class SyncTransformer extends Transformer {
 
   @override
   Future<String> transformRequest(RequestOptions options) async {
-    final dynamic data = options.data ?? '';
+    final Object data = options.data ?? '';
     if (data is! String && Transformer.isJsonMimeType(options.contentType)) {
       return jsonEncodeCallback(data);
-    } else if (data is Map<String, dynamic>) {
-      return Transformer.urlEncodeMap(data, options.listFormat);
+    } else if (data is Map) {
+      if (data is Map<String, dynamic>) {
+        return Transformer.urlEncodeMap(data, options.listFormat);
+      }
+      debugLog(
+        'The data is a type of `Map` (${data.runtimeType}), '
+        'but the transformer can only encode `Map<String, dynamic>`.\n'
+        'If you are writing maps using `{}`, '
+        'consider writing `<String, dynamic>{}`.',
+        StackTrace.current,
+      );
+      return data.toString();
     } else {
       return data.toString();
     }
@@ -48,9 +59,8 @@ class SyncTransformer extends Transformer {
       return responseBody;
     }
 
-    final showDownloadProgress = options.onReceiveProgress != null;
     final int totalLength;
-    if (showDownloadProgress) {
+    if (options.onReceiveProgress != null) {
       totalLength = int.parse(
         responseBody.headers[Headers.contentLengthHeader]?.first ?? '-1',
       );
@@ -58,27 +68,15 @@ class SyncTransformer extends Transformer {
       totalLength = 0;
     }
 
-    int received = 0;
-    final stream = responseBody.stream.transform<Uint8List>(
-      StreamTransformer.fromHandlers(
-        handleData: (data, sink) {
-          sink.add(data);
-          if (showDownloadProgress) {
-            received += data.length;
-            options.onReceiveProgress?.call(received, totalLength);
-          }
-        },
-      ),
-    );
-
     final streamCompleter = Completer<void>();
     int finalLength = 0;
     // Keep references to the data chunks and concatenate them later.
     final chunks = <Uint8List>[];
-    final subscription = stream.listen(
-      (chunk) {
+    final subscription = responseBody.stream.listen(
+      (Uint8List chunk) {
         finalLength += chunk.length;
         chunks.add(chunk);
+        options.onReceiveProgress?.call(finalLength, totalLength);
       },
       onError: (Object error, StackTrace stackTrace) {
         streamCompleter.completeError(error, stackTrace);
@@ -111,11 +109,17 @@ class SyncTransformer extends Transformer {
     );
     final String? response;
     if (options.responseDecoder != null) {
-      response = options.responseDecoder!(
+      final decodeResponse = options.responseDecoder!(
         responseBytes,
         options,
         responseBody..stream = Stream.empty(),
       );
+
+      if (decodeResponse is Future) {
+        response = await decodeResponse;
+      } else {
+        response = decodeResponse;
+      }
     } else if (!isJsonContent || responseBytes.isNotEmpty) {
       response = utf8.decode(responseBytes, allowMalformed: true);
     } else {