+++ dio/test/dio_mixin_test.dart
@@ -2,6 +2,11 @@ import 'package:dio/dio.dart';
 import 'package:test/test.dart';
 
 void main() {
+  test('not thrown for implements', () {
+    expect(_TestDioMixin().interceptors, isA<Interceptors>());
+    expect(_TestDioMixinExtends().interceptors, isA<Interceptors>());
+  });
+
   test('assureResponse', () {
     final requestOptions = RequestOptions(path: '');
     final untypedResponse = Response<dynamic>(
@@ -26,3 +31,5 @@ void main() {
 }
 
 class _TestDioMixin extends DioMixin implements Dio {}
+
+class _TestDioMixinExtends extends DioMixin implements Dio {}