+++ dio/test/download_test.dart
@@ -9,20 +9,38 @@ import 'mock/adapters.dart';
 import 'utils.dart';
 
 void main() {
+  late Directory tmp;
+
+  setUpAll(() {
+    tmp = Directory.systemTemp.createTempSync('dio_test_');
+    addTearDown(() {
+      tmp.deleteSync(recursive: true);
+    });
+  });
+
   setUp(startServer);
   tearDown(stopServer);
+
+  test('download does not change the response type', () async {
+    final savePath = p.join(tmp.path, 'download0.md');
+
+    final dio = Dio()..options.baseUrl = serverUrl.toString();
+    final options = Options(responseType: ResponseType.plain);
+    await dio.download('/download', savePath, options: options);
+    expect(options.responseType, ResponseType.plain);
+  });
+
   test('download1', () async {
-    const savePath = 'test/_download_test.md';
+    final savePath = p.join(tmp.path, 'download1.md');
     final dio = Dio()..options.baseUrl = serverUrl.toString();
     await dio.download('/download', savePath);
 
     final f = File(savePath);
     expect(f.readAsStringSync(), equals('I am a text file'));
-    f.deleteSync(recursive: false);
   });
 
   test('download2', () async {
-    const savePath = 'test/_download_test.md';
+    final savePath = p.join(tmp.path, 'download2.md');
     final dio = Dio()..options.baseUrl = serverUrl.toString();
     await dio.downloadUri(
       serverUrl.replace(path: '/download'),
@@ -31,11 +49,10 @@ void main() {
 
     final f = File(savePath);
     expect(f.readAsStringSync(), equals('I am a text file'));
-    f.deleteSync(recursive: false);
   });
 
   test('download error', () async {
-    const savePath = 'test/_download_test.md';
+    final savePath = p.join(tmp.path, 'download_error.md');
     final dio = Dio()..options.baseUrl = serverUrl.toString();
     Response response = await dio
         .download('/error', savePath)
@@ -52,23 +69,35 @@ void main() {
   });
 
   test('download timeout', () async {
-    const savePath = 'test/_download_test.md';
-    final dio = Dio(
-      BaseOptions(
-        receiveTimeout: Duration(milliseconds: 1),
-        baseUrl: serverUrl.toString(),
+    final dio = Dio();
+    final timeoutMatcher = allOf([
+      throwsA(isA<DioException>()),
+      throwsA(
+        predicate<DioException>(
+          (e) => e.type == DioExceptionType.receiveTimeout,
+        ),
       ),
+    ]);
+    await expectLater(
+      dio.downloadUri(
+        Uri.parse('$serverUrl/download').replace(
+          queryParameters: {'count': '3', 'gap': '2'},
+        ),
+        p.join(tmp.path, 'download_timeout.md'),
+        options: Options(receiveTimeout: Duration(seconds: 1)),
+      ),
+      timeoutMatcher,
     );
-    expect(
-      dio
-          .download('/download', savePath)
-          .catchError((e) => throw (e as DioException).type),
-      throwsA(DioExceptionType.receiveTimeout),
+    // Throws nothing if it constantly gets response bytes.
+    await dio.download(
+      'https://github.com/cfug/flutter.cn/archive/refs/heads/main.zip',
+      p.join(tmp.path, 'main.zip'),
+      options: Options(receiveTimeout: Duration(seconds: 1)),
     );
   });
 
   test('download cancellation', () async {
-    const savePath = 'test/_download_test.md';
+    final savePath = p.join(tmp.path, 'download_cancellation.md');
     final cancelToken = CancelToken();
     Future.delayed(Duration(milliseconds: 100), () {
       cancelToken.cancel();
@@ -86,7 +115,7 @@ void main() {
   });
 
   test('delete on error', () async {
-    const savePath = 'test/_download_test.md';
+    final savePath = p.join(tmp.path, 'delete_on_error.md');
     final f = File(savePath)..createSync(recursive: true);
     expect(f.existsSync(), isTrue);
 
@@ -106,7 +135,7 @@ void main() {
   });
 
   test('delete on cancel', () async {
-    const savePath = 'test/_download_test.md';
+    final savePath = p.join(tmp.path, 'delete_on_cancel.md');
     final f = File(savePath)..createSync(recursive: true);
     expect(f.existsSync(), isTrue);
 
@@ -129,7 +158,7 @@ void main() {
   });
 
   test('`savePath` types', () async {
-    final testPath = p.join(Directory.systemTemp.path, 'dio', 'testPath');
+    final testPath = p.join(tmp.path, 'savePath');
 
     final dio = Dio()
       ..options.baseUrl = EchoAdapter.mockBase