+++ plugins/cookie_manager/test/basic_test.dart
@@ -1,3 +1,4 @@
+@TestOn('vm')
 import 'package:cookie_jar/cookie_jar.dart';
 import 'package:dio/dio.dart';
 import 'package:dio_cookie_manager/dio_cookie_manager.dart';
@@ -7,12 +8,8 @@ void main() {
   test('cookie-jar', () async {
     final dio = Dio();
     final cookieJar = CookieJar();
-    dio.interceptors
-      ..add(CookieManager(cookieJar))
-      ..add(LogInterceptor());
+    dio.interceptors.add(CookieManager(cookieJar));
     await dio.get('https://pub.dev/');
-    // Print cookies
-    print(cookieJar.loadForRequest(Uri.parse('https://pub.dev/')));
     // second request with the cookie
     await dio.get('https://pub.dev/');
   });