+++ plugins/cookie_manager/test/cookies_test.dart
@@ -1,3 +1,4 @@
+@TestOn('vm')
 import 'dart:io';
 import 'dart:typed_data';