+++ plugins/http2_adapter/lib/src/connection_manager_imp.dart
@@ -77,15 +77,16 @@ class _ConnectionManager implements ConnectionManager {
       onClientCreate!(uri, clientConfig);
     }
 
-    late final SecureSocket socket;
-
+    // Allow [Socket] for non-TLS connections
+    // or [SecureSocket] for TLS connections.
+    late final Socket socket;
     try {
       socket = await _createSocket(uri, options, clientConfig);
     } on SocketException catch (e) {
       if (e.osError == null) {
         if (e.message.contains('timed out')) {
           throw DioException.connectionTimeout(
-            timeout: options.connectTimeout!,
+            timeout: options.connectTimeout ?? Duration.zero,
             requestOptions: options,
           );
         }
@@ -94,8 +95,10 @@ class _ConnectionManager implements ConnectionManager {
     }
 
     if (clientConfig.validateCertificate != null) {
+      final certificate =
+          socket is SecureSocket ? socket.peerCertificate : null;
       final isCertApproved = clientConfig.validateCertificate!(
-        socket.peerCertificate,
+        certificate,
         uri.host,
         uri.port,
       );
@@ -103,7 +106,7 @@ class _ConnectionManager implements ConnectionManager {
         throw DioException(
           requestOptions: options,
           type: DioExceptionType.badCertificate,
-          error: socket.peerCertificate,
+          error: certificate,
           message: 'The certificate of the response is not approved.',
         );
       }
@@ -119,7 +122,6 @@ class _ConnectionManager implements ConnectionManager {
       }
     };
 
-    //
     transportState.delayClose(
       _closed ? Duration(milliseconds: 50) : _idleTimeout,
       () {
@@ -130,16 +132,28 @@ class _ConnectionManager implements ConnectionManager {
     return transportState;
   }
 
-  Future<SecureSocket> _createSocket(
+  Future<Socket> _createSocket(
     Uri target,
     RequestOptions options,
     ClientSetting clientConfig,
   ) async {
-    if (clientConfig.proxy == null) {
+    final timeout = (options.connectTimeout ?? Duration.zero) > Duration.zero
+        ? options.connectTimeout!
+        : null;
+    final proxy = clientConfig.proxy;
+
+    if (proxy == null) {
+      if (target.scheme != 'https') {
+        return Socket.connect(
+          target.host,
+          target.port,
+          timeout: timeout,
+        );
+      }
       return SecureSocket.connect(
         target.host,
         target.port,
-        timeout: options.connectTimeout,
+        timeout: timeout,
         context: clientConfig.context,
         onBadCertificate: clientConfig.onBadCertificate,
         supportedProtocols: ['h2'],
@@ -147,13 +161,12 @@ class _ConnectionManager implements ConnectionManager {
     }
 
     final proxySocket = await Socket.connect(
-      clientConfig.proxy!.host,
-      clientConfig.proxy!.port,
-      timeout: options.connectTimeout,
+      proxy.host,
+      proxy.port,
+      timeout: timeout,
     );
 
-    final String credentialsProxy =
-        base64Encode(utf8.encode(clientConfig.proxy!.userInfo));
+    final String credentialsProxy = base64Encode(utf8.encode(proxy.userInfo));
 
     // Create http tunnel proxy https://www.ietf.org/rfc/rfc2817.txt