+++ plugins/http2_adapter/test/http2_test.dart
@@ -1,12 +1,18 @@
+@TestOn('vm')
 import 'package:dio/dio.dart';
 import 'package:dio_http2_adapter/dio_http2_adapter.dart';
 import 'package:test/test.dart';
 
 void main() {
+  test('works with non-TLS requests', () async {
+    final dio = Dio()..httpClientAdapter = Http2Adapter(ConnectionManager());
+    await dio.get('http://flutter.cn/');
+    await dio.get('https://flutter.cn/non-exist-destination');
+  });
+
   test('adds one to input values', () async {
     final dio = Dio()
-      ..options.baseUrl = 'https://pub.dev/'
-      ..interceptors.add(LogInterceptor())
+      ..options.baseUrl = 'https://httpbun.com/'
       ..httpClientAdapter = Http2Adapter(
         ConnectionManager(
           idleTimeout: Duration(milliseconds: 10),
@@ -15,7 +21,7 @@ void main() {
       );
 
     Response<String> response;
-    response = await dio.get('?xx=6');
+    response = await dio.get('get');
     assert(response.statusCode == 200);
     response = await dio.get(
       'nkjnjknjn.html',
@@ -37,18 +43,22 @@ void main() {
     expect(res.data.toString(), contains('TEST'));
   });
 
-  test('request with payload via proxy', () async {
-    final dio = Dio()
-      ..options.baseUrl = 'https://httpbun.com/'
-      ..httpClientAdapter = Http2Adapter(ConnectionManager(
-        idleTimeout: Duration(milliseconds: 10),
-        onClientCreate: (uri, settings) =>
-            settings.proxy = Uri.parse('http://localhost:3128'),
-      ));
+  test(
+    'request with payload via proxy',
+    () async {
+      final dio = Dio()
+        ..options.baseUrl = 'https://httpbun.com/'
+        ..httpClientAdapter = Http2Adapter(ConnectionManager(
+          idleTimeout: Duration(milliseconds: 10),
+          onClientCreate: (uri, settings) =>
+              settings.proxy = Uri.parse('http://localhost:3128'),
+        ));
 
-    final res = await dio.post('post', data: 'TEST');
-    expect(res.data.toString(), contains('TEST'));
-  });
+      final res = await dio.post('post', data: 'TEST');
+      expect(res.data.toString(), contains('TEST'));
+    },
+    tags: ['proxy'],
+  );
 
   test('request without network and restore', () async {
     bool needProxy = true;
@@ -116,5 +126,39 @@ void main() {
     );
 
     await dio.get('/drip?delay=1&numbytes=1');
+  });
+
+  test('request with redirect', () async {
+    final dio = Dio()
+      ..options.baseUrl = 'https://httpbun.com/'
+      ..httpClientAdapter = Http2Adapter(ConnectionManager());
+
+    final res = await dio.get('absolute-redirect/2');
+    expect(res.statusCode, 200);
+  });
+
+  test('header value types implicit support', () async {
+    final dio = Dio()
+      ..options.baseUrl = 'https://httpbun.com/'
+      ..httpClientAdapter = Http2Adapter(ConnectionManager());
+
+    final res = await dio.post(
+      'post',
+      data: 'TEST',
+      options: Options(
+        headers: {
+          'ListKey': ['1', '2'],
+          'StringKey': '1',
+          'NumKey': 2,
+          'BooleanKey': false,
+        },
+      ),
+    );
+    final content = res.data.toString();
+    expect(content, contains('TEST'));
+    expect(content, contains('Listkey: 1, 2'));
+    expect(content, contains('Stringkey: 1'));
+    expect(content, contains('Numkey: 2'));
+    expect(content, contains('Booleankey: false'));
   });
 }